requirements.txt
appdirs==1.4.0
click==6.7
cookies==2.2.1
coverage==4.3.4
flake8==3.2.1
mccabe==0.5.3
packaging==16.8
py==1.4.32
pycodestyle==2.2.0
pyflakes==1.3.0
pyparsing==2.1.10
pytest==3.0.5
pytest-cov==2.4.0
PyYAML==3.12
requests==2.12.5
responses==0.5.1
semantic-version==2.6.0
six==1.10.0
terminaltables==3.1.0
click-didyoumean==0.0.3
click-repl==0.1.1
